creates routecontainer to hold routes and api/nav functions below App.tsx
